\name{mqqcause-package}
\alias{mqqcause-package}
\docType{package}
\title{mqqcause: Multivariate Quantile-on-Quantile Granger Causality}
\description{Implements bivariate Quantile-on-Quantile Granger causality
and the multivariate (conditional) version with a vector of moderators
and optional interaction terms. A Sup-Wald summary across the
(theta, tau) grid is also provided (Troster, 2018).}
\author{Dr Merwan Roudane \email{merwanroudane920@gmail.com}\cr
  GitHub: \url{https://github.com/merwanroudane/qqcaus}}
\keyword{package}
